Add initial data section
