+++ _content/tour/algorithms-data.article
@@ -39,8 +39,24 @@ Hash table
 
 ** List
 
+Double linked list with Go
+
+.play algorithms/data/list/list.go
+
 ** Queue
+
+Implementing a Queue with Go
+
+.play algorithms/data/queue/queue.go
 
 ** Stack
 
+Implementing a Stack with Go
+
+.play algorithms/data/stack/stack.go
+
 ** Tree
+
+Implementing a binary tree with Go
+
+.play algorithms/data/tree/binary/example/main.go
