+++ _content/tour/algorithms/data/list/list.go
@@ -1,13 +1,20 @@
+//go:build OMIT
+// +build OMIT
+
 // All material is licensed under the Apache License Version 2.0, January 2004
 // http://www.apache.org/licenses/LICENSE-2.0
 
 // Package list implements of a doubly link list in Go.
-package list
+package main
 
 import (
 	"fmt"
 	"strings"
 )
+
+func main() {
+
+}
 
 // Node represents the data being stored.
 type Node struct {
