+++ _content/tour/algorithms/data/stack/stack.go
@@ -1,10 +1,17 @@
+//go:build OMIT
+// +build OMIT
+
 // All material is licensed under the Apache License Version 2.0, January 2004
 // http://www.apache.org/licenses/LICENSE-2.0
 
 // Package stack asks the student to implement a stack in Go.
-package stack
+package main
 
 import "errors"
+
+func main() {
+
+}
 
 // Data represents what is being stored on the stack.
 type Data struct {
